+++ javascript/node/selenium-webdriver/executors.js
@@ -16,45 +16,32 @@
 // under the License.
 
 /**
- * @fileoverview Various utilities for working with
- * {@link webdriver.CommandExecutor} implementations.
+ * @fileoverview Various utilities for working with {@link ./command.Executor}
+ * implementations.
  */
+
+ 'use strict';
 
 var HttpClient = require('./http').HttpClient,
     HttpExecutor = require('./http').Executor,
     promise = require('./lib/_base').require('webdriver.promise');
 
-
-
-/**
- * Wraps a promised {@link webdriver.CommandExecutor}, ensuring no commands
- * are executed until the wrapped executor has been fully built.
- * @param {!webdriver.promise.Promise.<!webdriver.CommandExecutor>} delegate The
- *     promised delegate.
- * @constructor
- * @implements {webdriver.CommandExecutor}
- */
-var DeferredExecutor = function(delegate) {
-
-  /** @override */
-  this.execute = function(command) {
-    return delegate.then(executor => executor.execute(command));
-  };
-};
+var DeferredExecutor = require('./lib/command').DeferredExecutor;
 
 
 // PUBLIC API
 
 
+/** @deprecated Use {@link ./lib/command.DeferredExecutor} instead. */
 exports.DeferredExecutor = DeferredExecutor;
 
 /**
  * Creates a command executor that uses WebDriver's JSON wire protocol.
- * @param {(string|!webdriver.promise.Promise.<string>)} url The server's URL,
+ * @param {(string|!webdriver.promise.Promise<string>)} url The server's URL,
  *     or a promise that will resolve to that URL.
  * @param {string=} opt_proxy (optional) The URL of the HTTP proxy for the
  *     client to use.
- * @returns {!webdriver.CommandExecutor} The new command executor.
+ * @returns {!./lib/command.Executor} The new command executor.
  */
 exports.createExecutor = function(url, opt_proxy) {
   return new DeferredExecutor(promise.when(url, function(url) {