+++ javascript/node/selenium-webdriver/index.js
@@ -20,9 +20,11 @@
  * public API and provides convenience assessors to certain sub-modules.
  */
 
-var base = require('./lib/_base');
 var builder = require('./builder');
 var error = require('./error');
+var base = require('./lib/_base');
+var command = require('./lib/command');
+var logging = require('./lib/logging');
 
 
 // NOTE: the remainder of this file is nasty and verbose, but the annotations
@@ -43,10 +45,6 @@ exports.By = require('./lib/by').By;
 
 /** @type {function(new: webdriver.Capabilities)} */
 exports.Capabilities = base.require('webdriver.Capabilities');
-
-
-/** @type {function(new: webdriver.Command)} */
-exports.Command = base.require('webdriver.Command');
 
 
 /** @type {function(new: webdriver.EventEmitter)} */
@@ -99,12 +97,6 @@ exports.WebElementPromise = base.require('webdriver.WebElementPromise');
 }));
 
 
-/** @type {webdriver.CommandName.} */
-(exports.__defineGetter__('CommandName', function() {
-  return base.require('webdriver.CommandName');
-}));
-
-
 /** @type {webdriver.Key.} */
 (exports.__defineGetter__('Key', function() {
   return base.require('webdriver.Key');
@@ -123,10 +115,7 @@ exports.WebElementPromise = base.require('webdriver.WebElementPromise');
 }));
 
 
-/** @type {webdriver.logging.} */
-(exports.__defineGetter__('logging', function() {
-  return base.exportPublicApi('webdriver.logging');
-}));
+exports.logging = logging;
 
 
 /** @type {webdriver.promise.} */