+++ javascript/node/selenium-webdriver/test/http/util_test.js
@@ -20,6 +20,7 @@
 var assert = require('assert'),
     http = require('http');
 
+var error = require('../../error');
 var util = require('../../http/util');
 
 describe('selenium-webdriver/http/util', function() {
@@ -76,8 +77,9 @@ describe('selenium-webdriver/http/util', function() {
       return util.getStatus(baseUrl).then(function() {
         throw Error('expected a failure');
       }, function(err) {
-        assert.equal(status, err.code);
-        assert.equal(value, err.message);
+        assert.ok(err instanceof error.WebDriverError);
+        assert.equal(err.code, 'unknown error');
+        assert.equal(err.message, value);
       });
     });
 
@@ -100,8 +102,9 @@ describe('selenium-webdriver/http/util', function() {
       return util.getStatus(baseUrl).then(function() {
         throw Error('expected a failure');
       }, function(err) {
-        assert.equal(status, err.code);
-        assert.equal(value, err.message);
+        assert.ok(err instanceof error.WebDriverError);
+        assert.equal(err.code, 'unknown error');
+        assert.equal(err.message, value);
       });
     });
   });